requirements.txt
langchain
langchain-community
langchain-core
streamlit
pandas
sentence-transformers
huggingface-hub
python-dotenv
requests
llama-cpp-python
psycopg2-binary
numpy
pgvector
nltk
spacy
textblob
scikit-learn
torch>=2.0.0
transformers>=4.30.0
